some basic event loop stuff going in bokeh
although this demo is cool, i think it shows the framework
is not suitable for what I am looking to do. Because it has
to close a loop to the server to get any interesting data,
any dynamically computed results have to be re-fetched via
some web interfaces (and possibly written to a file? although
I suppose the tornado framework could serve from memory too,
but then we'd have to shuffle pixels into some container format
like PNG and shove it over HTTP to render on the screen).

Bokeh is good for browsing data, I suspect it's not the tool
I want to use for the intial analysis...still, should keep
this branch around, because I think once we get to building
a browser for the data it seems like it could be an interesting
way to add annotations and so forth.