make some more todo notes
probably before going into aligned/focused we should try to adjust
global brightness variation across an image. there is a gradient of
intensity on each tile, due to the position of the light. This
can be compensated for I think, but, need to think about how to do it.
Time to read up...