add a first-take at an auto-stitching algorithm & cleanup new schema routine
This algorithm works but is a bit slow. The basic idea is to search
along one axis for the minimum MSE between a reference and moving
image, and then search along another axis for a minimum MSE.

generally, the algorithm is producing robust results as long as the
search window is big enough. Probably some "quality" metric has to
be applied to the image first to make sure it's actually good enough
to process (e.g. the table wasn't shaking, wasn't out of focus).