seam blending is now working!
finally figured out how to use the damn detail library in cv2.
had to split out a small test stub routine to really play with things
because the documentation isn't very clear on how this all works.

stitching and averaging of multiple images now happens after an
alignment has been found. So the total workflow is like this:

1. load a new schema, everything is unaligned
2. run an alignemnt; fixup any mis-aligned tiles
3. render a final image, which uses the base alignments as
a QC check and reference alignment for image duplicates. also
does blending to reduce seams.

Probably should add an explicit button for alignment pass, instead
of just having it be a hidden keyboard macro.