yet another big refactor
this modularizes the main loop of the stitcher and tracks the
state in an object.

mostly the benefit of this refactor is better isolation of
different stages of stitching. Before it was getting pretty
complicated and re-used variable names was leading to subtle
bugs. By splitting everything into separate fuction calls
it's clear where we forgot to initialize variables for a given
step.

It also has some added benefit of a cleaner call API, and
also the control flow of the central loop is really more clearly
around the user interaction and what to do about the stitching
results, rather than a mix of DSP algorithms and UI flow.

Finally, a bunch of variable names and really poorly thought
out vector transformations were simplified hopefully making future
maintenance easier(?)